[C8] Split fillet_curved_weld.go by responsibility (#2222)
fillet_curved_weld.go was 522 LOC, over the 500-line ceiling. Split it
along its three responsibilities so each file states one thing:

- fillet_curved_weld.go (273): weld routing + arm gathering/normalization
  and the result-face orchestration (weldCurvedArmOrFloor,
  assembleCurvedArmBody, trihedralCornerBody, curvedWeldFaces).
- fillet_curved_weld_armrails.go (170): the trimmed arm face and its four
  boundary rails — host contact, setback great-arc, far cross-section runout.
- fillet_curved_weld_corner.go (109): the corner sphere spherical-triangle
  cap and its setback-arc chaining.

Behaviour-preserving: every function body is byte-identical; nothing added
or deleted, only moved. The arm-rails file is named _armrails (not _arm) so
Go does not build-constrain it to GOARCH=arm. Fillet corpus and archguard
pass before and after with no golden updates. Net delta of recognizers /
tolerance constants / fallback sites / type assertions: 0.

Part of #2183 (C8: structure/size limits). Closes #2222.

(cherry picked from commit e153e3bff89ef98beb787414f2fcb3984a31e942)
