+++ kernel/ops/fillet_curved_weld_corner.go
@@ -1,0 +1,109 @@
+// SPDX-License-Identifier: GPL-2.0-only
+
+package ops
+
+import (
+	"oblikovati.org/kernel/geom"
+	"oblikovati.org/kernel/topo"
+	"oblikovati.org/math"
+)
+
+// Curved-fillet weld — the CORNER CAP case (M48 #2222 split of fillet_curved_weld.go). The corner
+// sphere's spherical-triangle patch that caps the three welded arms: the surface is validated through
+// the RailLoop blend engine (analyticSphere / coons4 / canal), while the octant boundary loop stays the
+// legacy chainSetbackArcs so B3 is byte-for-byte. The three setback great-arcs are each shared with an
+// arm, so they chain head-to-tail into the closed spherical triangle.
+
+// curvedCornerFace emits the corner patch: the SURFACE is validated through the RailLoop engine
+// (extractCurvedCorner→resolveBlend — analyticSphere wins the octant, coons4 the degenerate 4-sided
+// fill), while the octant's boundary LOOP stays the legacy chainSetbackArcs so B3 is byte-for-byte
+// (ADR-2 Step 1 strangler; sphere loop-collapse is a gated follow-up). Falls back to curvedSphereFace
+// wholesale on any engine decline (do-no-harm). The Kind-gate is load-bearing: only BlendKindSphere
+// (via the legacy loop), BlendKindCoons4, and the M6' BlendKindCanal are admitted; any other tier
+// (e.g. tri3) is NOT a valid curved corner and falls back rather than shipping a wrong 3-sided corner.
+func curvedCornerFace(w cornerWeld, sphere geom.Sphere, arms []edgeFillet, res Resolution) (filletFace, bool) {
+	loop, ok := extractCurvedCorner(w, arms, res)
+	if !ok {
+		return curvedSphereFace(w, sphere) // extractor declined — legacy octant path
+	}
+	patch, ok := resolveBlend(loop, res)
+	if !ok {
+		return curvedSphereFace(w, sphere)
+	}
+	switch patch.Kind {
+	case BlendKindSphere:
+		return curvedSphereFace(w, sphere) // octant: engine-validated surface == sphere, KEEP legacy loop (ADR-2 Step 1)
+	case BlendKindCoons4:
+		return patchToFilletFace(patch, topo.Lineage{}), true // degenerate 4-sided fill: take the engine's loops
+	case BlendKindCanal:
+		return patchToFilletFace(patch, topo.Lineage{}), true // M6' rolling-ball canal: same handoff as coons4 (surface + received-rail loops)
+	default:
+		return curvedSphereFace(w, sphere) // any other tier (e.g. tri3) is NOT a valid curved corner — do-no-harm
+	}
+}
+
+// curvedSphereFace builds the corner sphere's spherical-triangle patch: the surface is the solved corner
+// sphere, the loop is the three setback great-arcs (each shared with an arm) chained into a closed ring.
+// It carries NO parent lineage by design: the corner sphere is generated by the trihedral VERTEX (where
+// the three filleted edges meet), not by any single edge, so no filletEdgeProvenance edge-name applies —
+// this mirrors the planar corner patch (spherePatchFace), which likewise emits no single-edge parent.
+func curvedSphereFace(w cornerWeld, sphere geom.Sphere) (filletFace, bool) {
+	segs, ok := chainSetbackArcs(w)
+	if !ok {
+		return filletFace{}, false
+	}
+	return filletFace{surface: sphere, loops: []filletLoop{loopFromSegs(segs)}}, true
+}
+
+// chainSetbackArcs collects each arm's setback great-arc and chains them head-to-tail into the closed
+// spherical triangle (the three rails share the host-tangent points pairwise, so they close into a ring).
+func chainSetbackArcs(w cornerWeld) ([]endSeg, bool) {
+	arcs := make([]endSeg, 0, len(w.arms))
+	for _, a := range w.arms {
+		rail, ok := curvedSetbackRail(w, a)
+		if !ok {
+			return nil, false
+		}
+		arcs = append(arcs, endSeg{from: rail.PointAt(0), to: rail.PointAt(1), curve: rail, mid: rail.PointAt(0.5), arc: true})
+	}
+	return chainEndSegs(arcs, railGreatCircleTol*w.radius)
+}
+
+// chainEndSegs orders a set of endSegs head-to-tail into a closed ring, reversing a seg when it is met
+// from its `to` end. Returns false if the chain cannot close (a rail endpoint has no continuation).
+func chainEndSegs(segs []endSeg, tol float64) ([]endSeg, bool) {
+	used := make([]bool, len(segs))
+	out := make([]endSeg, 0, len(segs))
+	cur := segs[0].from
+	for range segs {
+		next, ok := takeEndSegFrom(segs, used, cur, tol)
+		if !ok {
+			return nil, false
+		}
+		out = append(out, next)
+		cur = next.to
+	}
+	if float64(cur.DistanceTo(out[0].from)) > tol {
+		return nil, false // chain did not return to its start — not a closed loop
+	}
+	return out, true
+}
+
+// takeEndSegFrom returns the first unused seg touching cur, oriented to start at cur (reversing it when
+// cur is its `to` end), and marks it used.
+func takeEndSegFrom(segs []endSeg, used []bool, cur math.Point3, tol float64) (endSeg, bool) {
+	for i := range segs {
+		if used[i] {
+			continue
+		}
+		if float64(segs[i].from.DistanceTo(cur)) <= tol {
+			used[i] = true
+			return segs[i], true
+		}
+		if float64(segs[i].to.DistanceTo(cur)) <= tol {
+			used[i] = true
+			return reverseEndSegs([]endSeg{segs[i]})[0], true
+		}
+	}
+	return endSeg{}, false
+}
